+++ bot.py
@@ -327,6 +327,8 @@ async def on_error(event, *args, **kwargs):
 bot.load_extension("src.discord.censor")
 bot.load_extension("src.discord.ping")
 bot.load_extension("src.discord.staffcommands")
+bot.load_extension("src.discord.staff.invitationals")
+bot.load_extension("src.discord.staff.censor")
 bot.load_extension("src.discord.embed")
 bot.load_extension("src.discord.membercommands")
 bot.load_extension("src.discord.devtools")
